Exception handling - update

+++ src/main/java/com/estudoswill/course/services/UserService.java
@@ -4,46 +4,58 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 
 import com.estudoswill.course.entities.User;
 import com.estudoswill.course.repositories.UserRepository;
 import com.estudoswill.course.services.exceptions.ResourceNotFoundException;
 
+import jakarta.persistence.EntityNotFoundException;
+
 // Os services funcionam como uma ponte entre os controlares e o Repository
 
 @Service
 public class UserService {
-	
+
 	@Autowired
 	private UserRepository userRepository;
-	
-	public List<User> findAll(){
+
+	public List<User> findAll() {
 		return userRepository.findAll();
 	}
-	
+
 	public User findById(Long id) {
 		Optional<User> obj = userRepository.findById(id);
-		
+
 		// Retorna objeto do tipo user de dentro do Optional
-		// Tenta fazer o get, mas se der problema ele lança oa exceção 
+		// Tenta fazer o get, mas se der problema ele lança oa exceção
 		return obj.orElseThrow(() -> new ResourceNotFoundException(id));
 	}
-	
+
 	// Função para salvar no banco de dados e retornar o Usuário salvo
 	public User insert(User obj) {
 		return userRepository.save(obj);
 	}
-	
+
 	public void delete(Long id) {
-		userRepository.deleteById(id);
+		try {
+			if (!userRepository.existsById(id)) {
+				throw new ResourceNotFoundException(id);
+			}
+			userRepository.deleteById(id);
+		} catch (ResourceNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
 	}
-	
+
 	public User update(Long id, User obj) {
-		User entity = userRepository.getReferenceById(id);
-		updateData(entity, obj);
-		return userRepository.save(entity);
+		try {
+			User entity = userRepository.getReferenceById(id);
+			updateData(entity, obj);
+			return userRepository.save(entity);
+		}catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException(id);
+		}
 	}
 
 	private void updateData(User entity, User obj) {
@@ -51,6 +63,5 @@ public class UserService {
 		entity.setEmail(obj.getEmail());
 		entity.setPhone(obj.getPhone());
 	}
-	
-	
+
 }
